experiments/requirements.txt
-e git+ssh://git@github.com/pedroilidio/bipartite_learn.git@5b58d6ae9b545b16d157eabe3959721171c9ef9a#egg=bipartite_learn&subdirectory=../../../../scripts/predictors/hypertrees
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Cython==0.29.33
fonttools==4.46.0
idna==3.6
imbalanced-learn==0.9.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
patsy==0.5.4
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.0
scikit-posthocs==0.8.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
statsmodels==0.14.0
tabulate==0.9.0
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.1.0
